update line chart
