+++ src/app/modules/charts/board/board.component.html
@@ -11,9 +11,7 @@
       </div>
 
       <div class="card" fxFlex="50">
-        <app-dotted-line-chart
-          title="OLE Over the last 5 year">
-        </app-dotted-line-chart>
+        <div kpiLineChart></div>
       </div>
     </div>
