+++ src/app/modules/charts/board/board.component.scss
@@ -14,5 +14,11 @@
     border-radius: 10px;
     border: 1px solid rgba(0, 0, 0, .22);
     margin: 8px;
+    background: rgba(0, 0, 0, .22);
+
+    div {
+      width: 100%;
+      height: 100%;
+    }
   }
 }
