--- src/app/modules/charts/dashed-line-chart/dashed-line-chart.component.html
@@ -1,8 +1,0 @@
-<div #dashedLineChart class="container">
-  <div class="header">
-    <div class="title">{{title}}</div>
-    <div class="subtitle">{{subtitle}}</div>
-  </div>
-
-  <div id="DASHED_LINE"></div>
-</div>
