--- src/app/modules/charts/dashed-line-chart/dashed-line-chart.component.scss
@@ -1,32 +1,0 @@
-.container {
-  width: 100%;
-  height: 100%;
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  overflow: auto;
-
-  .header {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-
-    .title {
-      font-size: 12px;
-      font-weight: 600;
-    }
-
-    .subtitle {
-      font-size: 11px;
-      font-weight: 400;
-      color: rgba(0, 0, 0, 0.68);
-    }
-  }
-
-  .chart {
-    overflow-y: auto;
-  }
-}
