--- src/app/modules/charts/dashed-line-chart/dashed-line-chart.component.spec.ts
@@ -1,25 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { DashedLineChartComponent } from './dashed-line-chart.component';
-
-describe('DashedLineChartComponent', () => {
-  let component: DashedLineChartComponent;
-  let fixture: ComponentFixture<DashedLineChartComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ DashedLineChartComponent ]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(DashedLineChartComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
