--- src/app/modules/charts/dashed-line-chart/mock.ts
@@ -1,5136 +1,0 @@
-export const MOCK_DATA = [
-  { date: '2019-01-04', value: 89},
-  { date: '2019-01-11', value: 96},
-  { date: '2019-01-18', value: 87},
-  { date: '2019-01-25', value: 99},
-  { date: '2019-02-04', value: 83},
-  { date: '2019-02-11', value: 93},
-  { date: '2019-02-18', value: 79},
-  { date: '2019-02-25', value: 94},
-  { date: '2019-03-4',  value: 89},
-  { date: '2019-03-11', value: 93},
-  { date: '2019-03-18', value: 81}
-];
-
-export const AAPL = [
-  {
-    date: '2007-04-23',
-    value: 93.24
-  },
-  {
-    date: '2007-04-24',
-    value: 95.35
-  },
-  {
-    date: '2007-04-25',
-    value: 98.84
-  },
-  {
-    date: '2007-04-26',
-    value: 99.92
-  },
-  {
-    date: '2007-04-29',
-    value: 99.8
-  },
-  {
-    date: '2007-05-01',
-    value: 99.47
-  },
-  {
-    date: '2007-05-02',
-    value: 100.39
-  },
-  {
-    date: '2007-05-03',
-    value: 100.4
-  },
-  {
-    date: '2007-05-04',
-    value: 100.81
-  },
-  {
-    date: '2007-05-07',
-    value: 103.92
-  },
-  {
-    date: '2007-05-08',
-    value: 105.06
-  },
-  {
-    date: '2007-05-09',
-    value: 106.88
-  },
-  {
-    date: '2007-05-09',
-    value: 107.34
-  },
-  {
-    date: '2007-05-10',
-    value: 108.74
-  },
-  {
-    date: '2007-05-13',
-    value: 109.36
-  },
-  {
-    date: '2007-05-14',
-    value: 107.52
-  },
-  {
-    date: '2007-05-15',
-    value: 107.34
-  },
-  {
-    date: '2007-05-16',
-    value: 109.44
-  },
-  {
-    date: '2007-05-17',
-    value: 110.02
-  },
-  {
-    date: '2007-05-20',
-    value: 111.98
-  },
-  {
-    date: '2007-05-21',
-    value: 113.54
-  },
-  {
-    date: '2007-05-22',
-    value: 112.89
-  },
-  {
-    date: '2007-05-23',
-    value: 110.69
-  },
-  {
-    date: '2007-05-24',
-    value: 113.62
-  },
-  {
-    date: '2007-05-28',
-    value: 114.35
-  },
-  {
-    date: '2007-05-29',
-    value: 118.77
-  },
-  {
-    date: '2007-05-30',
-    value: 121.19
-  },
-  {
-    date: '2007-06-01',
-    value: 118.4
-  },
-  {
-    date: '2007-06-04',
-    value: 121.33
-  },
-  {
-    date: '2007-06-05',
-    value: 122.67
-  },
-  {
-    date: '2007-06-06',
-    value: 123.64
-  },
-  {
-    date: '2007-06-07',
-    value: 124.07
-  },
-  {
-    date: '2007-06-08',
-    value: 124.49
-  },
-  {
-    date: '2007-06-10',
-    value: 120.19
-  },
-  {
-    date: '2007-06-11',
-    value: 120.38
-  },
-  {
-    date: '2007-06-12',
-    value: 117.5
-  },
-  {
-    date: '2007-06-13',
-    value: 118.75
-  },
-  {
-    date: '2007-06-14',
-    value: 120.5
-  },
-  {
-    date: '2007-06-17',
-    value: 125.09
-  },
-  {
-    date: '2007-06-18',
-    value: 123.66
-  },
-  {
-    date: '2007-06-19',
-    value: 121.55
-  },
-  {
-    date: '2007-06-20',
-    value: 123.9
-  },
-  {
-    date: '2007-06-21',
-    value: 123
-  },
-  {
-    date: '2007-06-24',
-    value: 122.34
-  },
-  {
-    date: '2007-06-25',
-    value: 119.65
-  },
-  {
-    date: '2007-06-26',
-    value: 121.89
-  },
-  {
-    date: '2007-06-27',
-    value: 120.56
-  },
-  {
-    date: '2007-06-28',
-    value: 122.04
-  },
-  {
-    date: '2007-07-02',
-    value: 121.26
-  },
-  {
-    date: '2007-07-03',
-    value: 127.17
-  },
-  {
-    date: '2007-07-05',
-    value: 132.75
-  },
-  {
-    date: '2007-07-06',
-    value: 132.3
-  },
-  {
-    date: '2007-07-09',
-    value: 130.33
-  },
-  {
-    date: '2007-07-09',
-    value: 132.35
-  },
-  {
-    date: '2007-07-10',
-    value: 132.39
-  },
-  {
-    date: '2007-07-11',
-    value: 134.07
-  },
-  {
-    date: '2007-07-12',
-    value: 137.73
-  },
-  {
-    date: '2007-07-15',
-    value: 138.1
-  },
-  {
-    date: '2007-07-16',
-    value: 138.91
-  },
-  {
-    date: '2007-07-17',
-    value: 138.12
-  },
-  {
-    date: '2007-07-18',
-    value: 140
-  },
-  {
-    date: '2007-07-19',
-    value: 143.75
-  },
-  {
-    date: '2007-07-22',
-    value: 143.7
-  },
-  {
-    date: '2007-07-23',
-    value: 134.89
-  },
-  {
-    date: '2007-07-24',
-    value: 137.26
-  },
-  {
-    date: '2007-07-25',
-    value: 146
-  },
-  {
-    date: '2007-07-26',
-    value: 143.85
-  },
-  {
-    date: '2007-07-29',
-    value: 141.43
-  },
-  {
-    date: '2007-07-30',
-    value: 131.76
-  },
-  {
-    date: '2007-08-01',
-    value: 135
-  },
-  {
-    date: '2007-08-02',
-    value: 136.49
-  },
-  {
-    date: '2007-08-03',
-    value: 131.85
-  },
-  {
-    date: '2007-08-06',
-    value: 135.25
-  },
-  {
-    date: '2007-08-07',
-    value: 135.03
-  },
-  {
-    date: '2007-08-08',
-    value: 134.01
-  },
-  {
-    date: '2007-08-09',
-    value: 126.39
-  },
-  {
-    date: '2007-08-09',
-    value: 125
-  },
-  {
-    date: '2007-08-12',
-    value: 127.79
-  },
-  {
-    date: '2007-08-13',
-    value: 124.03
-  },
-  {
-    date: '2007-08-14',
-    value: 119.9
-  },
-  {
-    date: '2007-08-15',
-    value: 117.05
-  },
-  {
-    date: '2007-08-16',
-    value: 122.06
-  },
-  {
-    date: '2007-08-19',
-    value: 122.22
-  },
-  {
-    date: '2007-08-20',
-    value: 127.57
-  },
-  {
-    date: '2007-08-21',
-    value: 132.51
-  },
-  {
-    date: '2007-08-22',
-    value: 131.07
-  },
-  {
-    date: '2007-08-23',
-    value: 135.3
-  },
-  {
-    date: '2007-08-26',
-    value: 132.25
-  },
-  {
-    date: '2007-08-27',
-    value: 126.82
-  },
-  {
-    date: '2007-08-28',
-    value: 134.08
-  },
-  {
-    date: '2007-08-29',
-    value: 136.25
-  },
-  {
-    date: '2007-08-30',
-    value: 138.48
-  },
-  {
-    date: '2007-09-04',
-    value: 144.16
-  },
-  {
-    date: '2007-09-05',
-    value: 136.76
-  },
-  {
-    date: '2007-09-06',
-    value: 135.01
-  },
-  {
-    date: '2007-09-07',
-    value: 131.77
-  },
-  {
-    date: '2007-09-09',
-    value: 136.71
-  },
-  {
-    date: '2007-09-10',
-    value: 135.49
-  },
-  {
-    date: '2007-09-11',
-    value: 136.85
-  },
-  {
-    date: '2007-09-12',
-    value: 137.2
-  },
-  {
-    date: '2007-09-13',
-    value: 138.81
-  },
-  {
-    date: '2007-09-16',
-    value: 138.41
-  },
-  {
-    date: '2007-09-17',
-    value: 140.92
-  },
-  {
-    date: '2007-09-18',
-    value: 140.77
-  },
-  {
-    date: '2007-09-19',
-    value: 140.31
-  },
-  {
-    date: '2007-09-20',
-    value: 144.15
-  },
-  {
-    date: '2007-09-23',
-    value: 148.28
-  },
-  {
-    date: '2007-09-24',
-    value: 153.18
-  },
-  {
-    date: '2007-09-25',
-    value: 152.77
-  },
-  {
-    date: '2007-09-26',
-    value: 154.5
-  },
-  {
-    date: '2007-09-27',
-    value: 153.47
-  },
-  {
-    date: '2007-10-01',
-    value: 156.34
-  },
-  {
-    date: '2007-10-02',
-    value: 158.45
-  },
-  {
-    date: '2007-10-03',
-    value: 157.92
-  },
-  {
-    date: '2007-10-04',
-    value: 156.24
-  },
-  {
-    date: '2007-10-05',
-    value: 161.45
-  },
-  {
-    date: '2007-10-08',
-    value: 167.91
-  },
-  {
-    date: '2007-10-09',
-    value: 167.86
-  },
-  {
-    date: '2007-10-09',
-    value: 166.79
-  },
-  {
-    date: '2007-10-10',
-    value: 162.23
-  },
-  {
-    date: '2007-10-11',
-    value: 167.25
-  },
-  {
-    date: '2007-10-14',
-    value: 166.98
-  },
-  {
-    date: '2007-10-15',
-    value: 169.58
-  },
-  {
-    date: '2007-10-16',
-    value: 172.75
-  },
-  {
-    date: '2007-10-17',
-    value: 173.5
-  },
-  {
-    date: '2007-10-18',
-    value: 170.42
-  },
-  {
-    date: '2007-10-21',
-    value: 174.36
-  },
-  {
-    date: '2007-10-22',
-    value: 186.16
-  },
-  {
-    date: '2007-10-23',
-    value: 185.93
-  },
-  {
-    date: '2007-10-24',
-    value: 182.78
-  },
-  {
-    date: '2007-10-25',
-    value: 184.7
-  },
-  {
-    date: '2007-10-28',
-    value: 185.09
-  },
-  {
-    date: '2007-10-29',
-    value: 187
-  },
-  {
-    date: '2007-10-30',
-    value: 189.95
-  },
-  {
-    date: '2007-11-01',
-    value: 187.44
-  },
-  {
-    date: '2007-11-02',
-    value: 187.87
-  },
-  {
-    date: '2007-11-05',
-    value: 186.18
-  },
-  {
-    date: '2007-11-06',
-    value: 191.79
-  },
-  {
-    date: '2007-11-07',
-    value: 186.3
-  },
-  {
-    date: '2007-11-08',
-    value: 175.47
-  },
-  {
-    date: '2007-11-09',
-    value: 165.37
-  },
-  {
-    date: '2007-11-11',
-    value: 153.76
-  },
-  {
-    date: '2007-11-12',
-    value: 169.96
-  },
-  {
-    date: '2007-11-13',
-    value: 166.11
-  },
-  {
-    date: '2007-11-14',
-    value: 164.3
-  },
-  {
-    date: '2007-11-15',
-    value: 166.39
-  },
-  {
-    date: '2007-11-18',
-    value: 163.95
-  },
-  {
-    date: '2007-11-19',
-    value: 168.85
-  },
-  {
-    date: '2007-11-20',
-    value: 168.46
-  },
-  {
-    date: '2007-11-22',
-    value: 171.54
-  },
-  {
-    date: '2007-11-25',
-    value: 172.54
-  },
-  {
-    date: '2007-11-26',
-    value: 174.81
-  },
-  {
-    date: '2007-11-27',
-    value: 180.22
-  },
-  {
-    date: '2007-11-28',
-    value: 184.29
-  },
-  {
-    date: '2007-11-29',
-    value: 182.22
-  },
-  {
-    date: '2007-12-03',
-    value: 178.86
-  },
-  {
-    date: '2007-12-04',
-    value: 179.81
-  },
-  {
-    date: '2007-12-05',
-    value: 185.5
-  },
-  {
-    date: '2007-12-06',
-    value: 189.95
-  },
-  {
-    date: '2007-12-07',
-    value: 194.3
-  },
-  {
-    date: '2007-12-09',
-    value: 194.21
-  },
-  {
-    date: '2007-12-10',
-    value: 188.54
-  },
-  {
-    date: '2007-12-11',
-    value: 190.86
-  },
-  {
-    date: '2007-12-12',
-    value: 191.83
-  },
-  {
-    date: '2007-12-13',
-    value: 190.39
-  },
-  {
-    date: '2007-12-16',
-    value: 184.4
-  },
-  {
-    date: '2007-12-17',
-    value: 182.98
-  },
-  {
-    date: '2007-12-18',
-    value: 183.12
-  },
-  {
-    date: '2007-12-19',
-    value: 187.21
-  },
-  {
-    date: '2007-12-20',
-    value: 193.91
-  },
-  {
-    date: '2007-12-23',
-    value: 198.8
-  },
-  {
-    date: '2007-12-25',
-    value: 198.95
-  },
-  {
-    date: '2007-12-26',
-    value: 198.57
-  },
-  {
-    date: '2007-12-27',
-    value: 199.83
-  },
-  {
-    date: '2007-12-30',
-    value: 198.08
-  },
-  {
-    date: '2008-01-02',
-    value: 194.84
-  },
-  {
-    date: '2008-01-03',
-    value: 194.93
-  },
-  {
-    date: '2008-01-04',
-    value: 180.05
-  },
-  {
-    date: '2008-01-07',
-    value: 177.64
-  },
-  {
-    date: '2008-01-08',
-    value: 171.25
-  },
-  {
-    date: '2008-01-09',
-    value: 179.4
-  },
-  {
-    date: '2008-01-09',
-    value: 178.02
-  },
-  {
-    date: '2008-01-10',
-    value: 172.69
-  },
-  {
-    date: '2008-01-13',
-    value: 178.78
-  },
-  {
-    date: '2008-01-14',
-    value: 169.04
-  },
-  {
-    date: '2008-01-15',
-    value: 159.64
-  },
-  {
-    date: '2008-01-16',
-    value: 160.89
-  },
-  {
-    date: '2008-01-17',
-    value: 161.36
-  },
-  {
-    date: '2008-01-21',
-    value: 155.64
-  },
-  {
-    date: '2008-01-22',
-    value: 139.07
-  },
-  {
-    date: '2008-01-23',
-    value: 135.6
-  },
-  {
-    date: '2008-01-24',
-    value: 130.01
-  },
-  {
-    date: '2008-01-27',
-    value: 130.01
-  },
-  {
-    date: '2008-01-28',
-    value: 131.54
-  },
-  {
-    date: '2008-01-29',
-    value: 132.18
-  },
-  {
-    date: '2008-01-30',
-    value: 135.36
-  },
-  {
-    date: '2008-02-01',
-    value: 133.75
-  },
-  {
-    date: '2008-02-04',
-    value: 131.65
-  },
-  {
-    date: '2008-02-05',
-    value: 129.36
-  },
-  {
-    date: '2008-02-06',
-    value: 122
-  },
-  {
-    date: '2008-02-07',
-    value: 121.24
-  },
-  {
-    date: '2008-02-08',
-    value: 125.48
-  },
-  {
-    date: '2008-02-10',
-    value: 129.45
-  },
-  {
-    date: '2008-02-11',
-    value: 124.86
-  },
-  {
-    date: '2008-02-12',
-    value: 129.4
-  },
-  {
-    date: '2008-02-13',
-    value: 127.46
-  },
-  {
-    date: '2008-02-14',
-    value: 124.63
-  },
-  {
-    date: '2008-02-18',
-    value: 122.18
-  },
-  {
-    date: '2008-02-19',
-    value: 123.82
-  },
-  {
-    date: '2008-02-20',
-    value: 121.54
-  },
-  {
-    date: '2008-02-21',
-    value: 119.46
-  },
-  {
-    date: '2008-02-24',
-    value: 119.74
-  },
-  {
-    date: '2008-02-25',
-    value: 119.15
-  },
-  {
-    date: '2008-02-26',
-    value: 122.96
-  },
-  {
-    date: '2008-02-27',
-    value: 129.91
-  },
-  {
-    date: '2008-02-28',
-    value: 125.02
-  },
-  {
-    date: '2008-03-03',
-    value: 121.73
-  },
-  {
-    date: '2008-03-04',
-    value: 124.62
-  },
-  {
-    date: '2008-03-05',
-    value: 124.49
-  },
-  {
-    date: '2008-03-06',
-    value: 120.93
-  },
-  {
-    date: '2008-03-07',
-    value: 122.25
-  },
-  {
-    date: '2008-03-09',
-    value: 119.69
-  },
-  {
-    date: '2008-03-10',
-    value: 127.35
-  },
-  {
-    date: '2008-03-11',
-    value: 126.03
-  },
-  {
-    date: '2008-03-12',
-    value: 127.94
-  },
-  {
-    date: '2008-03-13',
-    value: 126.61
-  },
-  {
-    date: '2008-03-16',
-    value: 126.73
-  },
-  {
-    date: '2008-03-17',
-    value: 132.82
-  },
-  {
-    date: '2008-03-18',
-    value: 129.67
-  },
-  {
-    date: '2008-03-19',
-    value: 133.27
-  },
-  {
-    date: '2008-03-23',
-    value: 139.53
-  },
-  {
-    date: '2008-03-24',
-    value: 140.98
-  },
-  {
-    date: '2008-03-25',
-    value: 145.06
-  },
-  {
-    date: '2008-03-26',
-    value: 140.25
-  },
-  {
-    date: '2008-03-27',
-    value: 143.01
-  },
-  {
-    date: '2008-03-30',
-    value: 143.5
-  },
-  {
-    date: '2008-04-01',
-    value: 149.53
-  },
-  {
-    date: '2008-04-02',
-    value: 147.49
-  },
-  {
-    date: '2008-04-03',
-    value: 151.61
-  },
-  {
-    date: '2008-04-04',
-    value: 153.08
-  },
-  {
-    date: '2008-04-07',
-    value: 155.89
-  },
-  {
-    date: '2008-04-08',
-    value: 152.84
-  },
-  {
-    date: '2008-04-09',
-    value: 151.44
-  },
-  {
-    date: '2008-04-09',
-    value: 154.55
-  },
-  {
-    date: '2008-04-10',
-    value: 147.14
-  },
-  {
-    date: '2008-04-13',
-    value: 147.78
-  },
-  {
-    date: '2008-04-14',
-    value: 148.38
-  },
-  {
-    date: '2008-04-15',
-    value: 153.7
-  },
-  {
-    date: '2008-04-16',
-    value: 154.49
-  },
-  {
-    date: '2008-04-17',
-    value: 161.04
-  },
-  {
-    date: '2008-04-20',
-    value: 168.16
-  },
-  {
-    date: '2008-04-21',
-    value: 160.2
-  },
-  {
-    date: '2008-04-22',
-    value: 162.89
-  },
-  {
-    date: '2008-04-23',
-    value: 168.94
-  },
-  {
-    date: '2008-04-24',
-    value: 169.73
-  },
-  {
-    date: '2008-04-27',
-    value: 172.24
-  },
-  {
-    date: '2008-04-28',
-    value: 175.05
-  },
-  {
-    date: '2008-04-29',
-    value: 173.95
-  },
-  {
-    date: '2008-05-01',
-    value: 180
-  },
-  {
-    date: '2008-05-02',
-    value: 180.94
-  },
-  {
-    date: '2008-05-05',
-    value: 184.73
-  },
-  {
-    date: '2008-05-06',
-    value: 186.66
-  },
-  {
-    date: '2008-05-07',
-    value: 182.59
-  },
-  {
-    date: '2008-05-08',
-    value: 185.06
-  },
-  {
-    date: '2008-05-09',
-    value: 183.45
-  },
-  {
-    date: '2008-05-11',
-    value: 188.16
-  },
-  {
-    date: '2008-05-12',
-    value: 189.96
-  },
-  {
-    date: '2008-05-13',
-    value: 186.26
-  },
-  {
-    date: '2008-05-14',
-    value: 189.73
-  },
-  {
-    date: '2008-05-15',
-    value: 187.62
-  },
-  {
-    date: '2008-05-18',
-    value: 183.6
-  },
-  {
-    date: '2008-05-19',
-    value: 185.9
-  },
-  {
-    date: '2008-05-20',
-    value: 178.19
-  },
-  {
-    date: '2008-05-21',
-    value: 177.05
-  },
-  {
-    date: '2008-05-22',
-    value: 181.17
-  },
-  {
-    date: '2008-05-26',
-    value: 186.43
-  },
-  {
-    date: '2008-05-27',
-    value: 187.01
-  },
-  {
-    date: '2008-05-28',
-    value: 186.69
-  },
-  {
-    date: '2008-05-29',
-    value: 188.75
-  },
-  {
-    date: '2008-06-02',
-    value: 186.1
-  },
-  {
-    date: '2008-06-03',
-    value: 185.37
-  },
-  {
-    date: '2008-06-04',
-    value: 185.19
-  },
-  {
-    date: '2008-06-05',
-    value: 189.43
-  },
-  {
-    date: '2008-06-06',
-    value: 185.64
-  },
-  {
-    date: '2008-06-09',
-    value: 181.61
-  },
-  {
-    date: '2008-06-09',
-    value: 185.64
-  },
-  {
-    date: '2008-06-10',
-    value: 180.81
-  },
-  {
-    date: '2008-06-11',
-    value: 173.26
-  },
-  {
-    date: '2008-06-12',
-    value: 172.37
-  },
-  {
-    date: '2008-06-15',
-    value: 176.84
-  },
-  {
-    date: '2008-06-16',
-    value: 181.43
-  },
-  {
-    date: '2008-06-17',
-    value: 178.75
-  },
-  {
-    date: '2008-06-18',
-    value: 180.9
-  },
-  {
-    date: '2008-06-19',
-    value: 175.27
-  },
-  {
-    date: '2008-06-22',
-    value: 173.16
-  },
-  {
-    date: '2008-06-23',
-    value: 173.25
-  },
-  {
-    date: '2008-06-24',
-    value: 177.39
-  },
-  {
-    date: '2008-06-25',
-    value: 168.26
-  },
-  {
-    date: '2008-06-26',
-    value: 170.09
-  },
-  {
-    date: '2008-06-29',
-    value: 167.44
-  },
-  {
-    date: '2008-07-01',
-    value: 174.68
-  },
-  {
-    date: '2008-07-02',
-    value: 168.18
-  },
-  {
-    date: '2008-07-03',
-    value: 170.12
-  },
-  {
-    date: '2008-07-07',
-    value: 175.16
-  },
-  {
-    date: '2008-07-08',
-    value: 179.55
-  },
-  {
-    date: '2008-07-09',
-    value: 174.25
-  },
-  {
-    date: '2008-07-09',
-    value: 176.63
-  },
-  {
-    date: '2008-07-10',
-    value: 172.58
-  },
-  {
-    date: '2008-07-13',
-    value: 173.88
-  },
-  {
-    date: '2008-07-14',
-    value: 169.64
-  },
-  {
-    date: '2008-07-15',
-    value: 172.81
-  },
-  {
-    date: '2008-07-16',
-    value: 171.81
-  },
-  {
-    date: '2008-07-17',
-    value: 165.15
-  },
-  {
-    date: '2008-07-20',
-    value: 166.29
-  },
-  {
-    date: '2008-07-21',
-    value: 162.02
-  },
-  {
-    date: '2008-07-22',
-    value: 166.26
-  },
-  {
-    date: '2008-07-23',
-    value: 159.03
-  },
-  {
-    date: '2008-07-24',
-    value: 162.12
-  },
-  {
-    date: '2008-07-27',
-    value: 154.4
-  },
-  {
-    date: '2008-07-28',
-    value: 157.08
-  },
-  {
-    date: '2008-07-29',
-    value: 159.88
-  },
-  {
-    date: '2008-07-30',
-    value: 158.95
-  },
-  {
-    date: '2008-08-01',
-    value: 156.66
-  },
-  {
-    date: '2008-08-04',
-    value: 153.23
-  },
-  {
-    date: '2008-08-05',
-    value: 160.64
-  },
-  {
-    date: '2008-08-06',
-    value: 164.19
-  },
-  {
-    date: '2008-08-07',
-    value: 163.57
-  },
-  {
-    date: '2008-08-08',
-    value: 169.55
-  },
-  {
-    date: '2008-08-10',
-    value: 173.56
-  },
-  {
-    date: '2008-08-11',
-    value: 176.73
-  },
-  {
-    date: '2008-08-12',
-    value: 179.3
-  },
-  {
-    date: '2008-08-13',
-    value: 179.32
-  },
-  {
-    date: '2008-08-14',
-    value: 175.74
-  },
-  {
-    date: '2008-08-17',
-    value: 175.39
-  },
-  {
-    date: '2008-08-18',
-    value: 173.53
-  },
-  {
-    date: '2008-08-19',
-    value: 175.84
-  },
-  {
-    date: '2008-08-20',
-    value: 174.29
-  },
-  {
-    date: '2008-08-21',
-    value: 176.79
-  },
-  {
-    date: '2008-08-24',
-    value: 172.55
-  },
-  {
-    date: '2008-08-25',
-    value: 173.64
-  },
-  {
-    date: '2008-08-26',
-    value: 174.67
-  },
-  {
-    date: '2008-08-27',
-    value: 173.74
-  },
-  {
-    date: '2008-08-28',
-    value: 169.53
-  },
-  {
-    date: '2008-09-02',
-    value: 166.19
-  },
-  {
-    date: '2008-09-03',
-    value: 166.96
-  },
-  {
-    date: '2008-09-04',
-    value: 161.22
-  },
-  {
-    date: '2008-09-05',
-    value: 160.18
-  },
-  {
-    date: '2008-09-08',
-    value: 157.92
-  },
-  {
-    date: '2008-09-09',
-    value: 151.68
-  },
-  {
-    date: '2008-09-09',
-    value: 151.61
-  },
-  {
-    date: '2008-09-10',
-    value: 152.65
-  },
-  {
-    date: '2008-09-11',
-    value: 148.94
-  },
-  {
-    date: '2008-09-14',
-    value: 140.36
-  },
-  {
-    date: '2008-09-15',
-    value: 139.88
-  },
-  {
-    date: '2008-09-16',
-    value: 127.83
-  },
-  {
-    date: '2008-09-17',
-    value: 134.09
-  },
-  {
-    date: '2008-09-18',
-    value: 140.91
-  },
-  {
-    date: '2008-09-21',
-    value: 131.05
-  },
-  {
-    date: '2008-09-22',
-    value: 126.84
-  },
-  {
-    date: '2008-09-23',
-    value: 128.71
-  },
-  {
-    date: '2008-09-24',
-    value: 131.93
-  },
-  {
-    date: '2008-09-25',
-    value: 128.24
-  },
-  {
-    date: '2008-09-28',
-    value: 105.26
-  },
-  {
-    date: '2008-09-29',
-    value: 113.66
-  },
-  {
-    date: '2008-10-01',
-    value: 109.12
-  },
-  {
-    date: '2008-10-02',
-    value: 100.1
-  },
-  {
-    date: '2008-10-03',
-    value: 97.07
-  },
-  {
-    date: '2008-10-06',
-    value: 98.14
-  },
-  {
-    date: '2008-10-07',
-    value: 89.16
-  },
-  {
-    date: '2008-10-08',
-    value: 89.79
-  },
-  {
-    date: '2008-10-09',
-    value: 88.74
-  },
-  {
-    date: '2008-10-09',
-    value: 96.8
-  },
-  {
-    date: '2008-10-12',
-    value: 110.26
-  },
-  {
-    date: '2008-10-13',
-    value: 104.08
-  },
-  {
-    date: '2008-10-14',
-    value: 97.95
-  },
-  {
-    date: '2008-10-15',
-    value: 101.89
-  },
-  {
-    date: '2008-10-16',
-    value: 97.4
-  },
-  {
-    date: '2008-10-19',
-    value: 98.44
-  },
-  {
-    date: '2008-10-20',
-    value: 91.49
-  },
-  {
-    date: '2008-10-21',
-    value: 96.87
-  },
-  {
-    date: '2008-10-22',
-    value: 98.23
-  },
-  {
-    date: '2008-10-23',
-    value: 96.38
-  },
-  {
-    date: '2008-10-26',
-    value: 92.09
-  },
-  {
-    date: '2008-10-27',
-    value: 99.91
-  },
-  {
-    date: '2008-10-28',
-    value: 104.55
-  },
-  {
-    date: '2008-10-29',
-    value: 111.04
-  },
-  {
-    date: '2008-10-30',
-    value: 107.59
-  },
-  {
-    date: '2008-11-03',
-    value: 106.96
-  },
-  {
-    date: '2008-11-04',
-    value: 110.99
-  },
-  {
-    date: '2008-11-05',
-    value: 103.3
-  },
-  {
-    date: '2008-11-06',
-    value: 99.1
-  },
-  {
-    date: '2008-11-07',
-    value: 98.24
-  },
-  {
-    date: '2008-11-09',
-    value: 95.88
-  },
-  {
-    date: '2008-11-10',
-    value: 94.77
-  },
-  {
-    date: '2008-11-11',
-    value: 90.12
-  },
-  {
-    date: '2008-11-12',
-    value: 96.44
-  },
-  {
-    date: '2008-11-13',
-    value: 90.24
-  },
-  {
-    date: '2008-11-16',
-    value: 88.14
-  },
-  {
-    date: '2008-11-17',
-    value: 89.91
-  },
-  {
-    date: '2008-11-18',
-    value: 86.29
-  },
-  {
-    date: '2008-11-19',
-    value: 80.49
-  },
-  {
-    date: '2008-11-20',
-    value: 82.58
-  },
-  {
-    date: '2008-11-23',
-    value: 92.95
-  },
-  {
-    date: '2008-11-24',
-    value: 90.8
-  },
-  {
-    date: '2008-11-25',
-    value: 95
-  },
-  {
-    date: '2008-11-26',
-    value: 95
-  },
-  {
-    date: '2008-11-27',
-    value: 92.67
-  },
-  {
-    date: '2008-12-01',
-    value: 88.93
-  },
-  {
-    date: '2008-12-02',
-    value: 92.47
-  },
-  {
-    date: '2008-12-03',
-    value: 95.9
-  },
-  {
-    date: '2008-12-04',
-    value: 91.41
-  },
-  {
-    date: '2008-12-05',
-    value: 94
-  },
-  {
-    date: '2008-12-08',
-    value: 99.72
-  },
-  {
-    date: '2008-12-09',
-    value: 100.06
-  },
-  {
-    date: '2008-12-09',
-    value: 98.21
-  },
-  {
-    date: '2008-12-10',
-    value: 95
-  },
-  {
-    date: '2008-12-11',
-    value: 98.27
-  },
-  {
-    date: '2008-12-14',
-    value: 94.75
-  },
-  {
-    date: '2008-12-15',
-    value: 95.43
-  },
-  {
-    date: '2008-12-16',
-    value: 89.16
-  },
-  {
-    date: '2008-12-17',
-    value: 89.43
-  },
-  {
-    date: '2008-12-18',
-    value: 90
-  },
-  {
-    date: '2008-12-21',
-    value: 85.74
-  },
-  {
-    date: '2008-12-22',
-    value: 86.38
-  },
-  {
-    date: '2008-12-23',
-    value: 85.04
-  },
-  {
-    date: '2008-12-24',
-    value: 85.04
-  },
-  {
-    date: '2008-12-25',
-    value: 85.81
-  },
-  {
-    date: '2008-12-28',
-    value: 86.61
-  },
-  {
-    date: '2008-12-29',
-    value: 86.29
-  },
-  {
-    date: '2008-12-30',
-    value: 85.35
-  },
-  {
-    date: '2009-01-01',
-    value: 85.35
-  },
-  {
-    date: '2009-01-02',
-    value: 90.75
-  },
-  {
-    date: '2009-01-05',
-    value: 94.58
-  },
-  {
-    date: '2009-01-06',
-    value: 93.02
-  },
-  {
-    date: '2009-01-07',
-    value: 91.01
-  },
-  {
-    date: '2009-01-08',
-    value: 92.7
-  },
-  {
-    date: '2009-01-09',
-    value: 90.58
-  },
-  {
-    date: '2009-01-11',
-    value: 88.66
-  },
-  {
-    date: '2009-01-12',
-    value: 87.71
-  },
-  {
-    date: '2009-01-13',
-    value: 85.33
-  },
-  {
-    date: '2009-01-14',
-    value: 83.38
-  },
-  {
-    date: '2009-01-15',
-    value: 82.33
-  },
-  {
-    date: '2009-01-19',
-    value: 78.2
-  },
-  {
-    date: '2009-01-20',
-    value: 82.83
-  },
-  {
-    date: '2009-01-21',
-    value: 88.36
-  },
-  {
-    date: '2009-01-22',
-    value: 88.36
-  },
-  {
-    date: '2009-01-25',
-    value: 89.64
-  },
-  {
-    date: '2009-01-26',
-    value: 90.73
-  },
-  {
-    date: '2009-01-27',
-    value: 94.2
-  },
-  {
-    date: '2009-01-28',
-    value: 93
-  },
-  {
-    date: '2009-01-29',
-    value: 90.13
-  },
-  {
-    date: '2009-02-02',
-    value: 91.51
-  },
-  {
-    date: '2009-02-03',
-    value: 92.98
-  },
-  {
-    date: '2009-02-04',
-    value: 93.55
-  },
-  {
-    date: '2009-02-05',
-    value: 96.46
-  },
-  {
-    date: '2009-02-06',
-    value: 99.72
-  },
-  {
-    date: '2009-02-09',
-    value: 102.51
-  },
-  {
-    date: '2009-02-09',
-    value: 97.83
-  },
-  {
-    date: '2009-02-10',
-    value: 96.82
-  },
-  {
-    date: '2009-02-11',
-    value: 99.27
-  },
-  {
-    date: '2009-02-12',
-    value: 99.16
-  },
-  {
-    date: '2009-02-16',
-    value: 94.53
-  },
-  {
-    date: '2009-02-17',
-    value: 94.37
-  },
-  {
-    date: '2009-02-18',
-    value: 90.64
-  },
-  {
-    date: '2009-02-19',
-    value: 91.2
-  },
-  {
-    date: '2009-02-22',
-    value: 86.95
-  },
-  {
-    date: '2009-02-23',
-    value: 90.25
-  },
-  {
-    date: '2009-02-24',
-    value: 91.16
-  },
-  {
-    date: '2009-02-25',
-    value: 89.19
-  },
-  {
-    date: '2009-02-26',
-    value: 89.31
-  },
-  {
-    date: '2009-03-02',
-    value: 87.94
-  },
-  {
-    date: '2009-03-03',
-    value: 88.37
-  },
-  {
-    date: '2009-03-04',
-    value: 91.17
-  },
-  {
-    date: '2009-03-05',
-    value: 88.84
-  },
-  {
-    date: '2009-03-06',
-    value: 85.3
-  },
-  {
-    date: '2009-03-09',
-    value: 83.11
-  },
-  {
-    date: '2009-03-09',
-    value: 88.63
-  },
-  {
-    date: '2009-03-10',
-    value: 92.68
-  },
-  {
-    date: '2009-03-11',
-    value: 96.35
-  },
-  {
-    date: '2009-03-12',
-    value: 95.93
-  },
-  {
-    date: '2009-03-15',
-    value: 95.42
-  },
-  {
-    date: '2009-03-16',
-    value: 99.66
-  },
-  {
-    date: '2009-03-17',
-    value: 101.52
-  },
-  {
-    date: '2009-03-18',
-    value: 101.62
-  },
-  {
-    date: '2009-03-19',
-    value: 101.59
-  },
-  {
-    date: '2009-03-22',
-    value: 107.66
-  },
-  {
-    date: '2009-03-23',
-    value: 106.5
-  },
-  {
-    date: '2009-03-24',
-    value: 106.49
-  },
-  {
-    date: '2009-03-25',
-    value: 109.87
-  },
-  {
-    date: '2009-03-26',
-    value: 106.85
-  },
-  {
-    date: '2009-03-29',
-    value: 104.49
-  },
-  {
-    date: '2009-03-30',
-    value: 105.12
-  },
-  {
-    date: '2009-04-01',
-    value: 108.69
-  },
-  {
-    date: '2009-04-02',
-    value: 112.71
-  },
-  {
-    date: '2009-04-03',
-    value: 115.99
-  },
-  {
-    date: '2009-04-06',
-    value: 118.45
-  },
-  {
-    date: '2009-04-07',
-    value: 115
-  },
-  {
-    date: '2009-04-08',
-    value: 116.32
-  },
-  {
-    date: '2009-04-09',
-    value: 119.57
-  },
-  {
-    date: '2009-04-09',
-    value: 119.57
-  },
-  {
-    date: '2009-04-12',
-    value: 120.22
-  },
-  {
-    date: '2009-04-13',
-    value: 118.31
-  },
-  {
-    date: '2009-04-14',
-    value: 117.64
-  },
-  {
-    date: '2009-04-15',
-    value: 121.45
-  },
-  {
-    date: '2009-04-16',
-    value: 123.42
-  },
-  {
-    date: '2009-04-19',
-    value: 120.5
-  },
-  {
-    date: '2009-04-20',
-    value: 121.76
-  },
-  {
-    date: '2009-04-21',
-    value: 121.51
-  },
-  {
-    date: '2009-04-22',
-    value: 125.4
-  },
-  {
-    date: '2009-04-23',
-    value: 123.9
-  },
-  {
-    date: '2009-04-26',
-    value: 124.73
-  },
-  {
-    date: '2009-04-27',
-    value: 123.9
-  },
-  {
-    date: '2009-04-28',
-    value: 125.14
-  },
-  {
-    date: '2009-04-29',
-    value: 125.83
-  },
-  {
-    date: '2009-05-01',
-    value: 127.24
-  },
-  {
-    date: '2009-05-04',
-    value: 132.07
-  },
-  {
-    date: '2009-05-05',
-    value: 132.71
-  },
-  {
-    date: '2009-05-06',
-    value: 132.5
-  },
-  {
-    date: '2009-05-07',
-    value: 129.06
-  },
-  {
-    date: '2009-05-08',
-    value: 129.19
-  },
-  {
-    date: '2009-05-10',
-    value: 129.57
-  },
-  {
-    date: '2009-05-11',
-    value: 124.42
-  },
-  {
-    date: '2009-05-12',
-    value: 119.49
-  },
-  {
-    date: '2009-05-13',
-    value: 122.95
-  },
-  {
-    date: '2009-05-14',
-    value: 122.42
-  },
-  {
-    date: '2009-05-17',
-    value: 126.65
-  },
-  {
-    date: '2009-05-18',
-    value: 127.45
-  },
-  {
-    date: '2009-05-19',
-    value: 125.87
-  },
-  {
-    date: '2009-05-20',
-    value: 124.18
-  },
-  {
-    date: '2009-05-21',
-    value: 122.5
-  },
-  {
-    date: '2009-05-25',
-    value: 130.78
-  },
-  {
-    date: '2009-05-26',
-    value: 133.05
-  },
-  {
-    date: '2009-05-27',
-    value: 135.07
-  },
-  {
-    date: '2009-05-28',
-    value: 135.81
-  },
-  {
-    date: '2009-06-01',
-    value: 139.35
-  },
-  {
-    date: '2009-06-02',
-    value: 139.49
-  },
-  {
-    date: '2009-06-03',
-    value: 140.95
-  },
-  {
-    date: '2009-06-04',
-    value: 143.74
-  },
-  {
-    date: '2009-06-05',
-    value: 144.67
-  },
-  {
-    date: '2009-06-08',
-    value: 143.85
-  },
-  {
-    date: '2009-06-09',
-    value: 142.72
-  },
-  {
-    date: '2009-06-09',
-    value: 140.25
-  },
-  {
-    date: '2009-06-10',
-    value: 139.95
-  },
-  {
-    date: '2009-06-11',
-    value: 136.97
-  },
-  {
-    date: '2009-06-14',
-    value: 136.09
-  },
-  {
-    date: '2009-06-15',
-    value: 136.35
-  },
-  {
-    date: '2009-06-16',
-    value: 135.58
-  },
-  {
-    date: '2009-06-17',
-    value: 135.88
-  },
-  {
-    date: '2009-06-18',
-    value: 139.48
-  },
-  {
-    date: '2009-06-21',
-    value: 137.37
-  },
-  {
-    date: '2009-06-22',
-    value: 134.01
-  },
-  {
-    date: '2009-06-23',
-    value: 136.22
-  },
-  {
-    date: '2009-06-24',
-    value: 139.86
-  },
-  {
-    date: '2009-06-25',
-    value: 142.44
-  },
-  {
-    date: '2009-06-28',
-    value: 141.97
-  },
-  {
-    date: '2009-06-29',
-    value: 142.43
-  },
-  {
-    date: '2009-07-01',
-    value: 142.83
-  },
-  {
-    date: '2009-07-02',
-    value: 140.02
-  },
-  {
-    date: '2009-07-03',
-    value: 140.02
-  },
-  {
-    date: '2009-07-06',
-    value: 138.61
-  },
-  {
-    date: '2009-07-07',
-    value: 135.4
-  },
-  {
-    date: '2009-07-08',
-    value: 137.22
-  },
-  {
-    date: '2009-07-09',
-    value: 136.36
-  },
-  {
-    date: '2009-07-09',
-    value: 138.52
-  },
-  {
-    date: '2009-07-12',
-    value: 142.34
-  },
-  {
-    date: '2009-07-13',
-    value: 142.27
-  },
-  {
-    date: '2009-07-14',
-    value: 146.88
-  },
-  {
-    date: '2009-07-15',
-    value: 147.52
-  },
-  {
-    date: '2009-07-16',
-    value: 151.75
-  },
-  {
-    date: '2009-07-19',
-    value: 152.91
-  },
-  {
-    date: '2009-07-20',
-    value: 151.51
-  },
-  {
-    date: '2009-07-21',
-    value: 156.74
-  },
-  {
-    date: '2009-07-22',
-    value: 157.82
-  },
-  {
-    date: '2009-07-23',
-    value: 159.99
-  },
-  {
-    date: '2009-07-26',
-    value: 160.1
-  },
-  {
-    date: '2009-07-27',
-    value: 160
-  },
-  {
-    date: '2009-07-28',
-    value: 160.03
-  },
-  {
-    date: '2009-07-29',
-    value: 162.79
-  },
-  {
-    date: '2009-07-30',
-    value: 163.39
-  },
-  {
-    date: '2009-08-03',
-    value: 166.43
-  },
-  {
-    date: '2009-08-04',
-    value: 165.55
-  },
-  {
-    date: '2009-08-05',
-    value: 165.11
-  },
-  {
-    date: '2009-08-06',
-    value: 163.91
-  },
-  {
-    date: '2009-08-07',
-    value: 165.51
-  },
-  {
-    date: '2009-08-09',
-    value: 164.72
-  },
-  {
-    date: '2009-08-11',
-    value: 165.31
-  },
-  {
-    date: '2009-08-12',
-    value: 168.42
-  },
-  {
-    date: '2009-08-13',
-    value: 166.78
-  },
-  {
-    date: '2009-08-16',
-    value: 159.59
-  },
-  {
-    date: '2009-08-17',
-    value: 164
-  },
-  {
-    date: '2009-08-18',
-    value: 164.6
-  },
-  {
-    date: '2009-08-19',
-    value: 166.33
-  },
-  {
-    date: '2009-08-20',
-    value: 169.22
-  },
-  {
-    date: '2009-08-23',
-    value: 169.06
-  },
-  {
-    date: '2009-08-24',
-    value: 169.4
-  },
-  {
-    date: '2009-08-25',
-    value: 167.41
-  },
-  {
-    date: '2009-08-26',
-    value: 169.45
-  },
-  {
-    date: '2009-08-27',
-    value: 170.05
-  },
-  {
-    date: '2009-08-30',
-    value: 168.21
-  },
-  {
-    date: '2009-09-01',
-    value: 165.3
-  },
-  {
-    date: '2009-09-02',
-    value: 165.18
-  },
-  {
-    date: '2009-09-03',
-    value: 166.55
-  },
-  {
-    date: '2009-09-04',
-    value: 170.31
-  },
-  {
-    date: '2009-09-08',
-    value: 172.93
-  },
-  {
-    date: '2009-09-09',
-    value: 171.14
-  },
-  {
-    date: '2009-09-09',
-    value: 172.56
-  },
-  {
-    date: '2009-09-10',
-    value: 172.16
-  },
-  {
-    date: '2009-09-13',
-    value: 173.72
-  },
-  {
-    date: '2009-09-14',
-    value: 175.16
-  },
-  {
-    date: '2009-09-15',
-    value: 181.87
-  },
-  {
-    date: '2009-09-16',
-    value: 184.55
-  },
-  {
-    date: '2009-09-17',
-    value: 185.02
-  },
-  {
-    date: '2009-09-20',
-    value: 184.02
-  },
-  {
-    date: '2009-09-21',
-    value: 184.48
-  },
-  {
-    date: '2009-09-22',
-    value: 185.5
-  },
-  {
-    date: '2009-09-23',
-    value: 183.82
-  },
-  {
-    date: '2009-09-24',
-    value: 182.37
-  },
-  {
-    date: '2009-09-27',
-    value: 186.15
-  },
-  {
-    date: '2009-09-28',
-    value: 185.38
-  },
-  {
-    date: '2009-09-29',
-    value: 185.35
-  },
-  {
-    date: '2009-10-01',
-    value: 180.86
-  },
-  {
-    date: '2009-10-02',
-    value: 184.9
-  },
-  {
-    date: '2009-10-05',
-    value: 186.02
-  },
-  {
-    date: '2009-10-06',
-    value: 190.01
-  },
-  {
-    date: '2009-10-07',
-    value: 190.25
-  },
-  {
-    date: '2009-10-08',
-    value: 189.27
-  },
-  {
-    date: '2009-10-09',
-    value: 190.47
-  },
-  {
-    date: '2009-10-11',
-    value: 190.81
-  },
-  {
-    date: '2009-10-12',
-    value: 190.02
-  },
-  {
-    date: '2009-10-13',
-    value: 191.29
-  },
-  {
-    date: '2009-10-14',
-    value: 190.56
-  },
-  {
-    date: '2009-10-15',
-    value: 188.05
-  },
-  {
-    date: '2009-10-18',
-    value: 189.86
-  },
-  {
-    date: '2009-10-19',
-    value: 198.76
-  },
-  {
-    date: '2009-10-20',
-    value: 204.92
-  },
-  {
-    date: '2009-10-21',
-    value: 205.2
-  },
-  {
-    date: '2009-10-22',
-    value: 203.94
-  },
-  {
-    date: '2009-10-25',
-    value: 202.48
-  },
-  {
-    date: '2009-10-26',
-    value: 197.37
-  },
-  {
-    date: '2009-10-27',
-    value: 192.4
-  },
-  {
-    date: '2009-10-28',
-    value: 196.35
-  },
-  {
-    date: '2009-10-29',
-    value: 188.5
-  },
-  {
-    date: '2009-11-02',
-    value: 189.31
-  },
-  {
-    date: '2009-11-03',
-    value: 188.75
-  },
-  {
-    date: '2009-11-04',
-    value: 190.81
-  },
-  {
-    date: '2009-11-05',
-    value: 194.03
-  },
-  {
-    date: '2009-11-06',
-    value: 194.34
-  },
-  {
-    date: '2009-11-09',
-    value: 201.46
-  },
-  {
-    date: '2009-11-09',
-    value: 202.98
-  },
-  {
-    date: '2009-11-10',
-    value: 203.25
-  },
-  {
-    date: '2009-11-11',
-    value: 201.99
-  },
-  {
-    date: '2009-11-12',
-    value: 204.45
-  },
-  {
-    date: '2009-11-15',
-    value: 206.63
-  },
-  {
-    date: '2009-11-16',
-    value: 207
-  },
-  {
-    date: '2009-11-17',
-    value: 205.96
-  },
-  {
-    date: '2009-11-18',
-    value: 200.51
-  },
-  {
-    date: '2009-11-19',
-    value: 199.92
-  },
-  {
-    date: '2009-11-22',
-    value: 205.88
-  },
-  {
-    date: '2009-11-23',
-    value: 204.44
-  },
-  {
-    date: '2009-11-24',
-    value: 204.19
-  },
-  {
-    date: '2009-11-25',
-    value: 204.19
-  },
-  {
-    date: '2009-11-26',
-    value: 200.59
-  },
-  {
-    date: '2009-11-29',
-    value: 199.91
-  },
-  {
-    date: '2009-12-01',
-    value: 196.97
-  },
-  {
-    date: '2009-12-02',
-    value: 196.23
-  },
-  {
-    date: '2009-12-03',
-    value: 196.48
-  },
-  {
-    date: '2009-12-04',
-    value: 193.32
-  },
-  {
-    date: '2009-12-07',
-    value: 188.95
-  },
-  {
-    date: '2009-12-08',
-    value: 189.87
-  },
-  {
-    date: '2009-12-09',
-    value: 197.8
-  },
-  {
-    date: '2009-12-09',
-    value: 196.43
-  },
-  {
-    date: '2009-12-10',
-    value: 194.67
-  },
-  {
-    date: '2009-12-13',
-    value: 196.98
-  },
-  {
-    date: '2009-12-14',
-    value: 194.17
-  },
-  {
-    date: '2009-12-15',
-    value: 195.03
-  },
-  {
-    date: '2009-12-16',
-    value: 191.86
-  },
-  {
-    date: '2009-12-17',
-    value: 195.43
-  },
-  {
-    date: '2009-12-20',
-    value: 198.23
-  },
-  {
-    date: '2009-12-21',
-    value: 200.36
-  },
-  {
-    date: '2009-12-22',
-    value: 202.1
-  },
-  {
-    date: '2009-12-23',
-    value: 209.04
-  },
-  {
-    date: '2009-12-24',
-    value: 209.04
-  },
-  {
-    date: '2009-12-27',
-    value: 211.61
-  },
-  {
-    date: '2009-12-28',
-    value: 209.1
-  },
-  {
-    date: '2009-12-29',
-    value: 211.64
-  },
-  {
-    date: '2009-12-30',
-    value: 210.73
-  },
-  {
-    date: '2010-01-01',
-    value: 210.73
-  },
-  {
-    date: '2010-01-04',
-    value: 214.01
-  },
-  {
-    date: '2010-01-05',
-    value: 214.38
-  },
-  {
-    date: '2010-01-06',
-    value: 210.97
-  },
-  {
-    date: '2010-01-07',
-    value: 210.58
-  },
-  {
-    date: '2010-01-08',
-    value: 211.98
-  },
-  {
-    date: '2010-01-10',
-    value: 210.11
-  },
-  {
-    date: '2010-01-11',
-    value: 207.72
-  },
-  {
-    date: '2010-01-12',
-    value: 210.65
-  },
-  {
-    date: '2010-01-13',
-    value: 209.43
-  },
-  {
-    date: '2010-01-14',
-    value: 205.93
-  },
-  {
-    date: '2010-01-17',
-    value: 205.93
-  },
-  {
-    date: '2010-01-18',
-    value: 215.04
-  },
-  {
-    date: '2010-01-19',
-    value: 211.72
-  },
-  {
-    date: '2010-01-20',
-    value: 208.07
-  },
-  {
-    date: '2010-01-21',
-    value: 197.75
-  },
-  {
-    date: '2010-01-24',
-    value: 203.08
-  },
-  {
-    date: '2010-01-25',
-    value: 205.94
-  },
-  {
-    date: '2010-01-26',
-    value: 207.88
-  },
-  {
-    date: '2010-01-27',
-    value: 199.29
-  },
-  {
-    date: '2010-01-28',
-    value: 192.06
-  },
-  {
-    date: '2010-02-01',
-    value: 194.73
-  },
-  {
-    date: '2010-02-02',
-    value: 195.86
-  },
-  {
-    date: '2010-02-03',
-    value: 199.23
-  },
-  {
-    date: '2010-02-04',
-    value: 192.05
-  },
-  {
-    date: '2010-02-05',
-    value: 195.46
-  },
-  {
-    date: '2010-02-08',
-    value: 194.12
-  },
-  {
-    date: '2010-02-09',
-    value: 196.19
-  },
-  {
-    date: '2010-02-09',
-    value: 195.12
-  },
-  {
-    date: '2010-02-10',
-    value: 198.67
-  },
-  {
-    date: '2010-02-11',
-    value: 200.38
-  },
-  {
-    date: '2010-02-14',
-    value: 200.38
-  },
-  {
-    date: '2010-02-15',
-    value: 203.4
-  },
-  {
-    date: '2010-02-16',
-    value: 202.55
-  },
-  {
-    date: '2010-02-17',
-    value: 202.93
-  },
-  {
-    date: '2010-02-18',
-    value: 201.67
-  },
-  {
-    date: '2010-02-21',
-    value: 200.42
-  },
-  {
-    date: '2010-02-22',
-    value: 197.06
-  },
-  {
-    date: '2010-02-23',
-    value: 200.66
-  },
-  {
-    date: '2010-02-24',
-    value: 202
-  },
-  {
-    date: '2010-02-25',
-    value: 204.62
-  },
-  {
-    date: '2010-03-01',
-    value: 208.99
-  },
-  {
-    date: '2010-03-02',
-    value: 208.85
-  },
-  {
-    date: '2010-03-03',
-    value: 209.33
-  },
-  {
-    date: '2010-03-04',
-    value: 210.71
-  },
-  {
-    date: '2010-03-05',
-    value: 218.95
-  },
-  {
-    date: '2010-03-08',
-    value: 219.08
-  },
-  {
-    date: '2010-03-09',
-    value: 223.02
-  },
-  {
-    date: '2010-03-09',
-    value: 224.84
-  },
-  {
-    date: '2010-03-10',
-    value: 225.5
-  },
-  {
-    date: '2010-03-11',
-    value: 226.6
-  },
-  {
-    date: '2010-03-14',
-    value: 223.84
-  },
-  {
-    date: '2010-03-15',
-    value: 224.45
-  },
-  {
-    date: '2010-03-16',
-    value: 224.12
-  },
-  {
-    date: '2010-03-17',
-    value: 224.65
-  },
-  {
-    date: '2010-03-18',
-    value: 222.25
-  },
-  {
-    date: '2010-03-21',
-    value: 224.75
-  },
-  {
-    date: '2010-03-22',
-    value: 228.36
-  },
-  {
-    date: '2010-03-23',
-    value: 229.37
-  },
-  {
-    date: '2010-03-24',
-    value: 226.65
-  },
-  {
-    date: '2010-03-25',
-    value: 230.9
-  },
-  {
-    date: '2010-03-28',
-    value: 232.39
-  },
-  {
-    date: '2010-03-29',
-    value: 235.84
-  },
-  {
-    date: '2010-03-30',
-    value: 235
-  },
-  {
-    date: '2010-04-01',
-    value: 235.97
-  },
-  {
-    date: '2010-04-02',
-    value: 235.97
-  },
-  {
-    date: '2010-04-05',
-    value: 238.49
-  },
-  {
-    date: '2010-04-06',
-    value: 239.54
-  },
-  {
-    date: '2010-04-07',
-    value: 240.6
-  },
-  {
-    date: '2010-04-08',
-    value: 239.95
-  },
-  {
-    date: '2010-04-09',
-    value: 241.79
-  },
-  {
-    date: '2010-04-11',
-    value: 242.29
-  },
-  {
-    date: '2010-04-12',
-    value: 242.43
-  },
-  {
-    date: '2010-04-13',
-    value: 245.69
-  },
-  {
-    date: '2010-04-14',
-    value: 248.92
-  },
-  {
-    date: '2010-04-15',
-    value: 247.4
-  },
-  {
-    date: '2010-04-18',
-    value: 247.07
-  },
-  {
-    date: '2010-04-19',
-    value: 244.59
-  },
-  {
-    date: '2010-04-20',
-    value: 259.22
-  },
-  {
-    date: '2010-04-21',
-    value: 266.47
-  },
-  {
-    date: '2010-04-22',
-    value: 270.83
-  },
-  {
-    date: '2010-04-25',
-    value: 269.5
-  },
-  {
-    date: '2010-04-26',
-    value: 262.04
-  },
-  {
-    date: '2010-04-27',
-    value: 261.6
-  },
-  {
-    date: '2010-04-28',
-    value: 268.64
-  },
-  {
-    date: '2010-04-29',
-    value: 261.09
-  },
-  {
-    date: '2010-05-03',
-    value: 266.35
-  },
-  {
-    date: '2010-05-04',
-    value: 258.68
-  },
-  {
-    date: '2010-05-05',
-    value: 255.98
-  },
-  {
-    date: '2010-05-06',
-    value: 246.25
-  },
-  {
-    date: '2010-05-07',
-    value: 235.86
-  },
-  {
-    date: '2010-05-09',
-    value: 253.99
-  },
-  {
-    date: '2010-05-10',
-    value: 256.52
-  },
-  {
-    date: '2010-05-11',
-    value: 262.09
-  },
-  {
-    date: '2010-05-12',
-    value: 258.36
-  },
-  {
-    date: '2010-05-13',
-    value: 253.82
-  },
-  {
-    date: '2010-05-16',
-    value: 254.22
-  },
-  {
-    date: '2010-05-17',
-    value: 252.36
-  },
-  {
-    date: '2010-05-18',
-    value: 248.34
-  },
-  {
-    date: '2010-05-19',
-    value: 237.76
-  },
-  {
-    date: '2010-05-20',
-    value: 242.32
-  },
-  {
-    date: '2010-05-23',
-    value: 246.76
-  },
-  {
-    date: '2010-05-24',
-    value: 245.22
-  },
-  {
-    date: '2010-05-25',
-    value: 244.11
-  },
-  {
-    date: '2010-05-26',
-    value: 253.35
-  },
-  {
-    date: '2010-05-27',
-    value: 256.88
-  },
-  {
-    date: '2010-05-30',
-    value: 256.88
-  },
-  {
-    date: '2010-06-01',
-    value: 260.83
-  },
-  {
-    date: '2010-06-02',
-    value: 263.95
-  },
-  {
-    date: '2010-06-03',
-    value: 263.12
-  },
-  {
-    date: '2010-06-04',
-    value: 255.96
-  },
-  {
-    date: '2010-06-07',
-    value: 250.94
-  },
-  {
-    date: '2010-06-08',
-    value: 249.33
-  },
-  {
-    date: '2010-06-09',
-    value: 243.2
-  },
-  {
-    date: '2010-06-09',
-    value: 250.51
-  },
-  {
-    date: '2010-06-10',
-    value: 253.51
-  },
-  {
-    date: '2010-06-13',
-    value: 254.28
-  },
-  {
-    date: '2010-06-14',
-    value: 259.69
-  },
-  {
-    date: '2010-06-15',
-    value: 267.25
-  },
-  {
-    date: '2010-06-16',
-    value: 271.87
-  },
-  {
-    date: '2010-06-17',
-    value: 274.07
-  },
-  {
-    date: '2010-06-20',
-    value: 270.17
-  },
-  {
-    date: '2010-06-21',
-    value: 273.85
-  },
-  {
-    date: '2010-06-22',
-    value: 270.97
-  },
-  {
-    date: '2010-06-23',
-    value: 269
-  },
-  {
-    date: '2010-06-24',
-    value: 266.7
-  },
-  {
-    date: '2010-06-27',
-    value: 268.3
-  },
-  {
-    date: '2010-06-28',
-    value: 256.17
-  },
-  {
-    date: '2010-06-29',
-    value: 251.53
-  },
-  {
-    date: '2010-07-01',
-    value: 248.48
-  },
-  {
-    date: '2010-07-02',
-    value: 246.94
-  },
-  {
-    date: '2010-07-05',
-    value: 246.94
-  },
-  {
-    date: '2010-07-06',
-    value: 248.63
-  },
-  {
-    date: '2010-07-07',
-    value: 258.66
-  },
-  {
-    date: '2010-07-08',
-    value: 258.09
-  },
-  {
-    date: '2010-07-09',
-    value: 259.62
-  },
-  {
-    date: '2010-07-11',
-    value: 257.28
-  },
-  {
-    date: '2010-07-12',
-    value: 251.8
-  },
-  {
-    date: '2010-07-13',
-    value: 252.73
-  },
-  {
-    date: '2010-07-14',
-    value: 251.45
-  },
-  {
-    date: '2010-07-15',
-    value: 249.9
-  },
-  {
-    date: '2010-07-18',
-    value: 245.58
-  },
-  {
-    date: '2010-07-19',
-    value: 251.89
-  },
-  {
-    date: '2010-07-20',
-    value: 254.24
-  },
-  {
-    date: '2010-07-21',
-    value: 259.02
-  },
-  {
-    date: '2010-07-22',
-    value: 259.94
-  },
-  {
-    date: '2010-07-25',
-    value: 259.28
-  },
-  {
-    date: '2010-07-26',
-    value: 264.08
-  },
-  {
-    date: '2010-07-27',
-    value: 260.96
-  },
-  {
-    date: '2010-07-28',
-    value: 258.11
-  },
-  {
-    date: '2010-07-29',
-    value: 257.25
-  },
-  {
-    date: '2010-08-02',
-    value: 261.85
-  },
-  {
-    date: '2010-08-03',
-    value: 261.93
-  },
-  {
-    date: '2010-08-04',
-    value: 262.98
-  },
-  {
-    date: '2010-08-05',
-    value: 261.7
-  },
-  {
-    date: '2010-08-06',
-    value: 260.09
-  },
-  {
-    date: '2010-08-09',
-    value: 261.75
-  },
-  {
-    date: '2010-08-09',
-    value: 259.41
-  },
-  {
-    date: '2010-08-10',
-    value: 250.19
-  },
-  {
-    date: '2010-08-11',
-    value: 251.79
-  },
-  {
-    date: '2010-08-12',
-    value: 249.1
-  },
-  {
-    date: '2010-08-15',
-    value: 247.64
-  },
-  {
-    date: '2010-08-16',
-    value: 251.97
-  },
-  {
-    date: '2010-08-17',
-    value: 253.07
-  },
-  {
-    date: '2010-08-18',
-    value: 249.88
-  },
-  {
-    date: '2010-08-19',
-    value: 249.64
-  },
-  {
-    date: '2010-08-22',
-    value: 245.8
-  },
-  {
-    date: '2010-08-23',
-    value: 239.93
-  },
-  {
-    date: '2010-08-24',
-    value: 242.89
-  },
-  {
-    date: '2010-08-25',
-    value: 240.28
-  },
-  {
-    date: '2010-08-26',
-    value: 241.62
-  },
-  {
-    date: '2010-08-29',
-    value: 242.5
-  },
-  {
-    date: '2010-08-30',
-    value: 243.1
-  },
-  {
-    date: '2010-09-01',
-    value: 250.33
-  },
-  {
-    date: '2010-09-02',
-    value: 252.17
-  },
-  {
-    date: '2010-09-03',
-    value: 258.77
-  },
-  {
-    date: '2010-09-06',
-    value: 258.77
-  },
-  {
-    date: '2010-09-07',
-    value: 257.81
-  },
-  {
-    date: '2010-09-08',
-    value: 262.92
-  },
-  {
-    date: '2010-09-09',
-    value: 263.07
-  },
-  {
-    date: '2010-09-09',
-    value: 263.41
-  },
-  {
-    date: '2010-09-12',
-    value: 267.04
-  },
-  {
-    date: '2010-09-13',
-    value: 268.06
-  },
-  {
-    date: '2010-09-14',
-    value: 270.22
-  },
-  {
-    date: '2010-09-15',
-    value: 276.57
-  },
-  {
-    date: '2010-09-16',
-    value: 275.37
-  },
-  {
-    date: '2010-09-19',
-    value: 283.23
-  },
-  {
-    date: '2010-09-20',
-    value: 283.77
-  },
-  {
-    date: '2010-09-21',
-    value: 287.75
-  },
-  {
-    date: '2010-09-22',
-    value: 288.92
-  },
-  {
-    date: '2010-09-23',
-    value: 292.32
-  },
-  {
-    date: '2010-09-26',
-    value: 291.16
-  },
-  {
-    date: '2010-09-27',
-    value: 286.86
-  },
-  {
-    date: '2010-09-28',
-    value: 287.37
-  },
-  {
-    date: '2010-09-29',
-    value: 283.75
-  },
-  {
-    date: '2010-10-01',
-    value: 282.52
-  },
-  {
-    date: '2010-10-04',
-    value: 278.64
-  },
-  {
-    date: '2010-10-05',
-    value: 288.94
-  },
-  {
-    date: '2010-10-06',
-    value: 289.19
-  },
-  {
-    date: '2010-10-07',
-    value: 289.22
-  },
-  {
-    date: '2010-10-08',
-    value: 294.07
-  },
-  {
-    date: '2010-10-10',
-    value: 295.36
-  },
-  {
-    date: '2010-10-11',
-    value: 298.54
-  },
-  {
-    date: '2010-10-12',
-    value: 300.14
-  },
-  {
-    date: '2010-10-13',
-    value: 302.31
-  },
-  {
-    date: '2010-10-14',
-    value: 314.74
-  },
-  {
-    date: '2010-10-17',
-    value: 318
-  },
-  {
-    date: '2010-10-18',
-    value: 309.49
-  },
-  {
-    date: '2010-10-19',
-    value: 310.53
-  },
-  {
-    date: '2010-10-20',
-    value: 309.52
-  },
-  {
-    date: '2010-10-21',
-    value: 307.47
-  },
-  {
-    date: '2010-10-24',
-    value: 308.84
-  },
-  {
-    date: '2010-10-25',
-    value: 308.05
-  },
-  {
-    date: '2010-10-26',
-    value: 307.83
-  },
-  {
-    date: '2010-10-27',
-    value: 305.24
-  },
-  {
-    date: '2010-10-28',
-    value: 300.98
-  },
-  {
-    date: '2010-11-01',
-    value: 304.18
-  },
-  {
-    date: '2010-11-02',
-    value: 309.36
-  },
-  {
-    date: '2010-11-03',
-    value: 312.8
-  },
-  {
-    date: '2010-11-04',
-    value: 318.27
-  },
-  {
-    date: '2010-11-05',
-    value: 317.13
-  },
-  {
-    date: '2010-11-08',
-    value: 318.62
-  },
-  {
-    date: '2010-11-09',
-    value: 316.08
-  },
-  {
-    date: '2010-11-09',
-    value: 318.03
-  },
-  {
-    date: '2010-11-10',
-    value: 316.66
-  },
-  {
-    date: '2010-11-11',
-    value: 308.03
-  },
-  {
-    date: '2010-11-14',
-    value: 307.04
-  },
-  {
-    date: '2010-11-15',
-    value: 301.59
-  },
-  {
-    date: '2010-11-16',
-    value: 300.5
-  },
-  {
-    date: '2010-11-17',
-    value: 308.43
-  },
-  {
-    date: '2010-11-18',
-    value: 306.73
-  },
-  {
-    date: '2010-11-21',
-    value: 313.36
-  },
-  {
-    date: '2010-11-22',
-    value: 308.73
-  },
-  {
-    date: '2010-11-23',
-    value: 314.8
-  },
-  {
-    date: '2010-11-25',
-    value: 315
-  },
-  {
-    date: '2010-11-28',
-    value: 316.87
-  },
-  {
-    date: '2010-11-29',
-    value: 311.15
-  },
-  {
-    date: '2010-12-01',
-    value: 316.4
-  },
-  {
-    date: '2010-12-02',
-    value: 318.15
-  },
-  {
-    date: '2010-12-03',
-    value: 317.44
-  },
-  {
-    date: '2010-12-06',
-    value: 320.15
-  },
-  {
-    date: '2010-12-07',
-    value: 318.21
-  },
-  {
-    date: '2010-12-08',
-    value: 321.01
-  },
-  {
-    date: '2010-12-09',
-    value: 319.76
-  },
-  {
-    date: '2010-12-09',
-    value: 320.56
-  },
-  {
-    date: '2010-12-12',
-    value: 321.67
-  },
-  {
-    date: '2010-12-13',
-    value: 320.29
-  },
-  {
-    date: '2010-12-14',
-    value: 320.36
-  },
-  {
-    date: '2010-12-15',
-    value: 321.25
-  },
-  {
-    date: '2010-12-16',
-    value: 320.61
-  },
-  {
-    date: '2010-12-19',
-    value: 322.21
-  },
-  {
-    date: '2010-12-20',
-    value: 324.2
-  },
-  {
-    date: '2010-12-21',
-    value: 325.16
-  },
-  {
-    date: '2010-12-22',
-    value: 323.6
-  },
-  {
-    date: '2010-12-26',
-    value: 324.68
-  },
-  {
-    date: '2010-12-27',
-    value: 325.47
-  },
-  {
-    date: '2010-12-28',
-    value: 325.29
-  },
-  {
-    date: '2010-12-29',
-    value: 323.66
-  },
-  {
-    date: '2010-12-30',
-    value: 322.56
-  },
-  {
-    date: '2011-01-03',
-    value: 329.57
-  },
-  {
-    date: '2011-01-04',
-    value: 331.29
-  },
-  {
-    date: '2011-01-05',
-    value: 334
-  },
-  {
-    date: '2011-01-06',
-    value: 333.73
-  },
-  {
-    date: '2011-01-07',
-    value: 336.12
-  },
-  {
-    date: '2011-01-09',
-    value: 342.46
-  },
-  {
-    date: '2011-01-10',
-    value: 341.64
-  },
-  {
-    date: '2011-01-11',
-    value: 344.42
-  },
-  {
-    date: '2011-01-12',
-    value: 345.68
-  },
-  {
-    date: '2011-01-13',
-    value: 348.48
-  },
-  {
-    date: '2011-01-17',
-    value: 340.65
-  },
-  {
-    date: '2011-01-18',
-    value: 338.84
-  },
-  {
-    date: '2011-01-19',
-    value: 332.68
-  },
-  {
-    date: '2011-01-20',
-    value: 326.72
-  },
-  {
-    date: '2011-01-23',
-    value: 337.45
-  },
-  {
-    date: '2011-01-24',
-    value: 341.4
-  },
-  {
-    date: '2011-01-25',
-    value: 343.85
-  },
-  {
-    date: '2011-01-26',
-    value: 343.21
-  },
-  {
-    date: '2011-01-27',
-    value: 336.1
-  },
-  {
-    date: '2011-01-30',
-    value: 339.32
-  },
-  {
-    date: '2011-02-01',
-    value: 345.03
-  },
-  {
-    date: '2011-02-02',
-    value: 344.32
-  },
-  {
-    date: '2011-02-03',
-    value: 343.44
-  },
-  {
-    date: '2011-02-04',
-    value: 346.5
-  },
-  {
-    date: '2011-02-07',
-    value: 351.88
-  },
-  {
-    date: '2011-02-08',
-    value: 355.2
-  },
-  {
-    date: '2011-02-09',
-    value: 358.16
-  },
-  {
-    date: '2011-02-09',
-    value: 354.54
-  },
-  {
-    date: '2011-02-10',
-    value: 356.85
-  },
-  {
-    date: '2011-02-13',
-    value: 359.18
-  },
-  {
-    date: '2011-02-14',
-    value: 359.9
-  },
-  {
-    date: '2011-02-15',
-    value: 363.13
-  },
-  {
-    date: '2011-02-16',
-    value: 358.3
-  },
-  {
-    date: '2011-02-17',
-    value: 350.56
-  },
-  {
-    date: '2011-02-21',
-    value: 338.61
-  },
-  {
-    date: '2011-02-22',
-    value: 342.62
-  },
-  {
-    date: '2011-02-23',
-    value: 342.88
-  },
-  {
-    date: '2011-02-24',
-    value: 348.16
-  },
-  {
-    date: '2011-02-27',
-    value: 353.21
-  },
-  {
-    date: '2011-03-01',
-    value: 349.31
-  },
-  {
-    date: '2011-03-02',
-    value: 352.12
-  },
-  {
-    date: '2011-03-03',
-    value: 359.56
-  },
-  {
-    date: '2011-03-04',
-    value: 360
-  },
-  {
-    date: '2011-03-07',
-    value: 355.36
-  },
-  {
-    date: '2011-03-08',
-    value: 355.76
-  },
-  {
-    date: '2011-03-09',
-    value: 352.47
-  },
-  {
-    date: '2011-03-09',
-    value: 346.67
-  },
-  {
-    date: '2011-03-10',
-    value: 351.99
-  },
-  {
-    date: '2011-03-13',
-    value: 353.56
-  },
-  {
-    date: '2011-03-14',
-    value: 345.43
-  },
-  {
-    date: '2011-03-15',
-    value: 330.01
-  },
-  {
-    date: '2011-03-16',
-    value: 334.64
-  },
-  {
-    date: '2011-03-17',
-    value: 330.67
-  },
-  {
-    date: '2011-03-20',
-    value: 339.3
-  },
-  {
-    date: '2011-03-21',
-    value: 341.2
-  },
-  {
-    date: '2011-03-22',
-    value: 339.19
-  },
-  {
-    date: '2011-03-23',
-    value: 344.97
-  },
-  {
-    date: '2011-03-24',
-    value: 351.54
-  },
-  {
-    date: '2011-03-27',
-    value: 350.44
-  },
-  {
-    date: '2011-03-28',
-    value: 350.96
-  },
-  {
-    date: '2011-03-29',
-    value: 348.63
-  },
-  {
-    date: '2011-03-30',
-    value: 348.51
-  },
-  {
-    date: '2011-04-01',
-    value: 344.56
-  },
-  {
-    date: '2011-04-04',
-    value: 341.19
-  },
-  {
-    date: '2011-04-05',
-    value: 338.89
-  },
-  {
-    date: '2011-04-06',
-    value: 338.04
-  },
-  {
-    date: '2011-04-07',
-    value: 338.08
-  },
-  {
-    date: '2011-04-08',
-    value: 335.06
-  },
-  {
-    date: '2011-04-10',
-    value: 330.8
-  },
-  {
-    date: '2011-04-11',
-    value: 332.4
-  },
-  {
-    date: '2011-04-12',
-    value: 336.13
-  },
-  {
-    date: '2011-04-13',
-    value: 332.42
-  },
-  {
-    date: '2011-04-14',
-    value: 327.46
-  },
-  {
-    date: '2011-04-17',
-    value: 331.85
-  },
-  {
-    date: '2011-04-18',
-    value: 337.86
-  },
-  {
-    date: '2011-04-19',
-    value: 342.41
-  },
-  {
-    date: '2011-04-20',
-    value: 350.7
-  },
-  {
-    date: '2011-04-24',
-    value: 353.01
-  },
-  {
-    date: '2011-04-25',
-    value: 350.42
-  },
-  {
-    date: '2011-04-26',
-    value: 350.15
-  },
-  {
-    date: '2011-04-27',
-    value: 346.75
-  },
-  {
-    date: '2011-04-28',
-    value: 350.13
-  },
-  {
-    date: '2011-05-02',
-    value: 346.28
-  },
-  {
-    date: '2011-05-03',
-    value: 348.2
-  },
-  {
-    date: '2011-05-04',
-    value: 349.57
-  },
-  {
-    date: '2011-05-05',
-    value: 346.75
-  },
-  {
-    date: '2011-05-06',
-    value: 346.66
-  },
-  {
-    date: '2011-05-09',
-    value: 347.6
-  },
-  {
-    date: '2011-05-09',
-    value: 349.45
-  },
-  {
-    date: '2011-05-10',
-    value: 347.23
-  },
-  {
-    date: '2011-05-11',
-    value: 346.57
-  },
-  {
-    date: '2011-05-12',
-    value: 340.5
-  },
-  {
-    date: '2011-05-15',
-    value: 333.3
-  },
-  {
-    date: '2011-05-16',
-    value: 336.14
-  },
-  {
-    date: '2011-05-17',
-    value: 339.87
-  },
-  {
-    date: '2011-05-18',
-    value: 340.53
-  },
-  {
-    date: '2011-05-19',
-    value: 335.22
-  },
-  {
-    date: '2011-05-22',
-    value: 334.4
-  },
-  {
-    date: '2011-05-23',
-    value: 332.19
-  },
-  {
-    date: '2011-05-24',
-    value: 336.78
-  },
-  {
-    date: '2011-05-25',
-    value: 335
-  },
-  {
-    date: '2011-05-26',
-    value: 337.41
-  },
-  {
-    date: '2011-05-30',
-    value: 347.83
-  },
-  {
-    date: '2011-06-01',
-    value: 345.51
-  },
-  {
-    date: '2011-06-02',
-    value: 346.1
-  },
-  {
-    date: '2011-06-03',
-    value: 343.44
-  },
-  {
-    date: '2011-06-06',
-    value: 338.04
-  },
-  {
-    date: '2011-06-07',
-    value: 332.04
-  },
-  {
-    date: '2011-06-08',
-    value: 332.24
-  },
-  {
-    date: '2011-06-09',
-    value: 331.49
-  },
-  {
-    date: '2011-06-09',
-    value: 325.9
-  },
-  {
-    date: '2011-06-12',
-    value: 326.6
-  },
-  {
-    date: '2011-06-13',
-    value: 332.44
-  },
-  {
-    date: '2011-06-14',
-    value: 326.75
-  },
-  {
-    date: '2011-06-15',
-    value: 325.16
-  },
-  {
-    date: '2011-06-16',
-    value: 320.26
-  },
-  {
-    date: '2011-06-19',
-    value: 315.32
-  },
-  {
-    date: '2011-06-20',
-    value: 325.3
-  },
-  {
-    date: '2011-06-21',
-    value: 322.61
-  },
-  {
-    date: '2011-06-22',
-    value: 331.23
-  },
-  {
-    date: '2011-06-23',
-    value: 326.35
-  },
-  {
-    date: '2011-06-26',
-    value: 332.04
-  },
-  {
-    date: '2011-06-27',
-    value: 335.26
-  },
-  {
-    date: '2011-06-28',
-    value: 334.04
-  },
-  {
-    date: '2011-06-29',
-    value: 335.67
-  },
-  {
-    date: '2011-07-01',
-    value: 343.26
-  },
-  {
-    date: '2011-07-05',
-    value: 349.43
-  },
-  {
-    date: '2011-07-06',
-    value: 351.76
-  },
-  {
-    date: '2011-07-07',
-    value: 357.2
-  },
-  {
-    date: '2011-07-08',
-    value: 359.71
-  },
-  {
-    date: '2011-07-10',
-    value: 354
-  },
-  {
-    date: '2011-07-11',
-    value: 353.75
-  },
-  {
-    date: '2011-07-12',
-    value: 358.02
-  },
-  {
-    date: '2011-07-13',
-    value: 357.77
-  },
-  {
-    date: '2011-07-14',
-    value: 364.92
-  },
-  {
-    date: '2011-07-17',
-    value: 373.8
-  },
-  {
-    date: '2011-07-18',
-    value: 376.85
-  },
-  {
-    date: '2011-07-19',
-    value: 386.9
-  },
-  {
-    date: '2011-07-20',
-    value: 387.29
-  },
-  {
-    date: '2011-07-21',
-    value: 393.3
-  },
-  {
-    date: '2011-07-24',
-    value: 398.5
-  },
-  {
-    date: '2011-07-25',
-    value: 403.41
-  },
-  {
-    date: '2011-07-26',
-    value: 392.59
-  },
-  {
-    date: '2011-07-27',
-    value: 391.82
-  },
-  {
-    date: '2011-07-28',
-    value: 390.48
-  },
-  {
-    date: '2011-08-01',
-    value: 396.75
-  },
-  {
-    date: '2011-08-02',
-    value: 388.91
-  },
-  {
-    date: '2011-08-03',
-    value: 392.57
-  },
-  {
-    date: '2011-08-04',
-    value: 377.37
-  },
-  {
-    date: '2011-08-05',
-    value: 373.62
-  },
-  {
-    date: '2011-08-08',
-    value: 353.21
-  },
-  {
-    date: '2011-08-09',
-    value: 374.01
-  },
-  {
-    date: '2011-08-09',
-    value: 363.69
-  },
-  {
-    date: '2011-08-10',
-    value: 373.7
-  },
-  {
-    date: '2011-08-11',
-    value: 376.99
-  },
-  {
-    date: '2011-08-14',
-    value: 383.41
-  },
-  {
-    date: '2011-08-15',
-    value: 380.48
-  },
-  {
-    date: '2011-08-16',
-    value: 380.44
-  },
-  {
-    date: '2011-08-17',
-    value: 366.05
-  },
-  {
-    date: '2011-08-18',
-    value: 356.03
-  },
-  {
-    date: '2011-08-21',
-    value: 356.44
-  },
-  {
-    date: '2011-08-22',
-    value: 373.6
-  },
-  {
-    date: '2011-08-23',
-    value: 376.18
-  },
-  {
-    date: '2011-08-24',
-    value: 373.72
-  },
-  {
-    date: '2011-08-25',
-    value: 383.58
-  },
-  {
-    date: '2011-08-28',
-    value: 389.97
-  },
-  {
-    date: '2011-08-29',
-    value: 389.99
-  },
-  {
-    date: '2011-08-30',
-    value: 384.83
-  },
-  {
-    date: '2011-09-01',
-    value: 381.03
-  },
-  {
-    date: '2011-09-02',
-    value: 374.05
-  },
-  {
-    date: '2011-09-06',
-    value: 379.74
-  },
-  {
-    date: '2011-09-07',
-    value: 383.93
-  },
-  {
-    date: '2011-09-08',
-    value: 384.14
-  },
-  {
-    date: '2011-09-09',
-    value: 377.48
-  },
-  {
-    date: '2011-09-11',
-    value: 379.94
-  },
-  {
-    date: '2011-09-12',
-    value: 384.62
-  },
-  {
-    date: '2011-09-13',
-    value: 389.3
-  },
-  {
-    date: '2011-09-14',
-    value: 392.96
-  },
-  {
-    date: '2011-09-15',
-    value: 400.5
-  },
-  {
-    date: '2011-09-18',
-    value: 411.63
-  },
-  {
-    date: '2011-09-19',
-    value: 413.45
-  },
-  {
-    date: '2011-09-20',
-    value: 412.14
-  },
-  {
-    date: '2011-09-21',
-    value: 401.82
-  },
-  {
-    date: '2011-09-22',
-    value: 404.3
-  },
-  {
-    date: '2011-09-25',
-    value: 403.17
-  },
-  {
-    date: '2011-09-26',
-    value: 399.26
-  },
-  {
-    date: '2011-09-27',
-    value: 397.01
-  },
-  {
-    date: '2011-09-28',
-    value: 390.57
-  },
-  {
-    date: '2011-09-29',
-    value: 381.32
-  },
-  {
-    date: '2011-10-03',
-    value: 374.6
-  },
-  {
-    date: '2011-10-04',
-    value: 372.5
-  },
-  {
-    date: '2011-10-05',
-    value: 378.25
-  },
-  {
-    date: '2011-10-06',
-    value: 377.37
-  },
-  {
-    date: '2011-10-07',
-    value: 369.8
-  },
-  {
-    date: '2011-10-09',
-    value: 388.81
-  },
-  {
-    date: '2011-10-10',
-    value: 400.29
-  },
-  {
-    date: '2011-10-11',
-    value: 402.19
-  },
-  {
-    date: '2011-10-12',
-    value: 408.43
-  },
-  {
-    date: '2011-10-13',
-    value: 422
-  },
-  {
-    date: '2011-10-16',
-    value: 419.99
-  },
-  {
-    date: '2011-10-17',
-    value: 422.24
-  },
-  {
-    date: '2011-10-18',
-    value: 398.62
-  },
-  {
-    date: '2011-10-19',
-    value: 395.31
-  },
-  {
-    date: '2011-10-20',
-    value: 392.87
-  },
-  {
-    date: '2011-10-23',
-    value: 405.77
-  },
-  {
-    date: '2011-10-24',
-    value: 397.77
-  },
-  {
-    date: '2011-10-25',
-    value: 400.6
-  },
-  {
-    date: '2011-10-26',
-    value: 404.69
-  },
-  {
-    date: '2011-10-27',
-    value: 404.95
-  },
-  {
-    date: '2011-10-30',
-    value: 404.78
-  },
-  {
-    date: '2011-11-01',
-    value: 396.51
-  },
-  {
-    date: '2011-11-02',
-    value: 397.41
-  },
-  {
-    date: '2011-11-03',
-    value: 403.07
-  },
-  {
-    date: '2011-11-04',
-    value: 400.24
-  },
-  {
-    date: '2011-11-07',
-    value: 399.73
-  },
-  {
-    date: '2011-11-08',
-    value: 406.23
-  },
-  {
-    date: '2011-11-09',
-    value: 395.28
-  },
-  {
-    date: '2011-11-09',
-    value: 385.22
-  },
-  {
-    date: '2011-11-10',
-    value: 384.62
-  },
-  {
-    date: '2011-11-13',
-    value: 379.26
-  },
-  {
-    date: '2011-11-14',
-    value: 388.83
-  },
-  {
-    date: '2011-11-15',
-    value: 384.77
-  },
-  {
-    date: '2011-11-16',
-    value: 377.41
-  },
-  {
-    date: '2011-11-17',
-    value: 374.94
-  },
-  {
-    date: '2011-11-20',
-    value: 369.01
-  },
-  {
-    date: '2011-11-21',
-    value: 376.51
-  },
-  {
-    date: '2011-11-22',
-    value: 366.99
-  },
-  {
-    date: '2011-11-24',
-    value: 363.57
-  },
-  {
-    date: '2011-11-27',
-    value: 376.12
-  },
-  {
-    date: '2011-11-28',
-    value: 373.2
-  },
-  {
-    date: '2011-11-29',
-    value: 382.2
-  },
-  {
-    date: '2011-12-01',
-    value: 387.93
-  },
-  {
-    date: '2011-12-02',
-    value: 389.7
-  },
-  {
-    date: '2011-12-05',
-    value: 393.01
-  },
-  {
-    date: '2011-12-06',
-    value: 390.95
-  },
-  {
-    date: '2011-12-07',
-    value: 389.09
-  },
-  {
-    date: '2011-12-08',
-    value: 390.66
-  },
-  {
-    date: '2011-12-09',
-    value: 393.62
-  },
-  {
-    date: '2011-12-11',
-    value: 391.84
-  },
-  {
-    date: '2011-12-12',
-    value: 388.81
-  },
-  {
-    date: '2011-12-13',
-    value: 380.19
-  },
-  {
-    date: '2011-12-14',
-    value: 378.94
-  },
-  {
-    date: '2011-12-15',
-    value: 381.02
-  },
-  {
-    date: '2011-12-18',
-    value: 382.21
-  },
-  {
-    date: '2011-12-19',
-    value: 395.95
-  },
-  {
-    date: '2011-12-20',
-    value: 396.44
-  },
-  {
-    date: '2011-12-21',
-    value: 398.55
-  },
-  {
-    date: '2011-12-22',
-    value: 403.43
-  },
-  {
-    date: '2011-12-26',
-    value: 406.53
-  },
-  {
-    date: '2011-12-27',
-    value: 402.64
-  },
-  {
-    date: '2011-12-28',
-    value: 405.12
-  },
-  {
-    date: '2011-12-29',
-    value: 405
-  },
-  {
-    date: '2012-01-03',
-    value: 411.23
-  },
-  {
-    date: '2012-01-04',
-    value: 413.44
-  },
-  {
-    date: '2012-01-05',
-    value: 418.03
-  },
-  {
-    date: '2012-01-06',
-    value: 422.4
-  },
-  {
-    date: '2012-01-09',
-    value: 421.73
-  },
-  {
-    date: '2012-01-09',
-    value: 423.24
-  },
-  {
-    date: '2012-01-10',
-    value: 422.55
-  },
-  {
-    date: '2012-01-11',
-    value: 421.39
-  },
-  {
-    date: '2012-01-12',
-    value: 419.81
-  },
-  {
-    date: '2012-01-16',
-    value: 424.7
-  },
-  {
-    date: '2012-01-17',
-    value: 429.11
-  },
-  {
-    date: '2012-01-18',
-    value: 427.75
-  },
-  {
-    date: '2012-01-19',
-    value: 420.3
-  },
-  {
-    date: '2012-01-22',
-    value: 427.41
-  },
-  {
-    date: '2012-01-23',
-    value: 420.41
-  },
-  {
-    date: '2012-01-24',
-    value: 446.66
-  },
-  {
-    date: '2012-01-25',
-    value: 444.63
-  },
-  {
-    date: '2012-01-26',
-    value: 447.28
-  },
-  {
-    date: '2012-01-29',
-    value: 453.01
-  },
-  {
-    date: '2012-01-30',
-    value: 456.48
-  },
-  {
-    date: '2012-02-01',
-    value: 456.19
-  },
-  {
-    date: '2012-02-02',
-    value: 455.12
-  },
-  {
-    date: '2012-02-03',
-    value: 459.68
-  },
-  {
-    date: '2012-02-06',
-    value: 463.97
-  },
-  {
-    date: '2012-02-07',
-    value: 468.83
-  },
-  {
-    date: '2012-02-08',
-    value: 476.68
-  },
-  {
-    date: '2012-02-09',
-    value: 493.17
-  },
-  {
-    date: '2012-02-09',
-    value: 493.42
-  },
-  {
-    date: '2012-02-12',
-    value: 502.6
-  },
-  {
-    date: '2012-02-13',
-    value: 509.46
-  },
-  {
-    date: '2012-02-14',
-    value: 497.67
-  },
-  {
-    date: '2012-02-15',
-    value: 502.21
-  },
-  {
-    date: '2012-02-16',
-    value: 502.12
-  },
-  {
-    date: '2012-02-20',
-    value: 514.85
-  },
-  {
-    date: '2012-02-21',
-    value: 513.04
-  },
-  {
-    date: '2012-02-22',
-    value: 516.39
-  },
-  {
-    date: '2012-02-23',
-    value: 522.41
-  },
-  {
-    date: '2012-02-26',
-    value: 525.76
-  },
-  {
-    date: '2012-02-27',
-    value: 535.41
-  },
-  {
-    date: '2012-02-28',
-    value: 542.44
-  },
-  {
-    date: '2012-03-01',
-    value: 544.47
-  },
-  {
-    date: '2012-03-02',
-    value: 545.18
-  },
-  {
-    date: '2012-03-05',
-    value: 533.16
-  },
-  {
-    date: '2012-03-06',
-    value: 530.26
-  },
-  {
-    date: '2012-03-07',
-    value: 530.69
-  },
-  {
-    date: '2012-03-08',
-    value: 541.99
-  },
-  {
-    date: '2012-03-09',
-    value: 545.17
-  },
-  {
-    date: '2012-03-11',
-    value: 552
-  },
-  {
-    date: '2012-03-12',
-    value: 568.1
-  },
-  {
-    date: '2012-03-13',
-    value: 589.58
-  },
-  {
-    date: '2012-03-14',
-    value: 585.56
-  },
-  {
-    date: '2012-03-15',
-    value: 585.57
-  },
-  {
-    date: '2012-03-18',
-    value: 601.1
-  },
-  {
-    date: '2012-03-19',
-    value: 605.96
-  },
-  {
-    date: '2012-03-20',
-    value: 602.5
-  },
-  {
-    date: '2012-03-21',
-    value: 599.34
-  },
-  {
-    date: '2012-03-22',
-    value: 596.05
-  },
-  {
-    date: '2012-03-25',
-    value: 606.98
-  },
-  {
-    date: '2012-03-26',
-    value: 614.48
-  },
-  {
-    date: '2012-03-27',
-    value: 617.62
-  },
-  {
-    date: '2012-03-28',
-    value: 609.86
-  },
-  {
-    date: '2012-03-29',
-    value: 599.55
-  },
-  {
-    date: '2012-04-02',
-    value: 618.63
-  },
-  {
-    date: '2012-04-03',
-    value: 629.32
-  },
-  {
-    date: '2012-04-04',
-    value: 624.31
-  },
-  {
-    date: '2012-04-05',
-    value: 633.68
-  },
-  {
-    date: '2012-04-09',
-    value: 636.23
-  },
-  {
-    date: '2012-04-09',
-    value: 628.44
-  },
-  {
-    date: '2012-04-10',
-    value: 626.2
-  },
-  {
-    date: '2012-04-11',
-    value: 622.77
-  },
-  {
-    date: '2012-04-12',
-    value: 605.23
-  },
-  {
-    date: '2012-04-15',
-    value: 580.13
-  },
-  {
-    date: '2012-04-16',
-    value: 609.7
-  },
-  {
-    date: '2012-04-17',
-    value: 608.34
-  },
-  {
-    date: '2012-04-18',
-    value: 587.44
-  },
-  {
-    date: '2012-04-19',
-    value: 572.98
-  },
-  {
-    date: '2012-04-22',
-    value: 571.7
-  },
-  {
-    date: '2012-04-23',
-    value: 560.28
-  },
-  {
-    date: '2012-04-24',
-    value: 610
-  },
-  {
-    date: '2012-04-25',
-    value: 607.7
-  },
-  {
-    date: '2012-04-26',
-    value: 603
-  },
-  {
-    date: '2012-04-29',
-    value: 583.98
-  },
-  {
-    date: '2012-05-01',
-    value: 582.13
-  }
-];
