--- src/app/modules/charts/dotted-line-chart/dotted-line-chart.component.html
@@ -1,8 +1,0 @@
-<div #dottedLineChart class="container">
-  <div class="header">
-    <div class="title">{{title}}</div>
-    <div class="subtitle">{{subtitle}}</div>
-  </div>
-
-  <div id="DOTTED_LINE_CHART"></div>
-</div>
