--- src/app/modules/charts/dotted-line-chart/mock.ts
@@ -1,14 +1,0 @@
-export const MOCK_DATA = [
-  { date: '2019-01-04', value: 89},
-  { date: '2019-01-11', value: 96},
-  { date: '2019-01-18', value: 87},
-  { date: '2019-01-25', value: 99},
-  { date: '2019-02-04', value: 83},
-  { date: '2019-02-11', value: 93},
-  { date: '2019-02-18', value: 79},
-  { date: '2019-02-25', value: 94},
-  { date: '2019-03-4',  value: 89},
-  { date: '2019-03-11', value: 93},
-  { date: '2019-03-18', value: 81}
-];
-
