+++ src/app/modules/charts/line-chart/line-chart.component.html
@@ -1,3 +1,3 @@
-<div class="container">
+<div #lineChartContainer class="container">
   <div id="LINE_CHART"></div>
 </div>
