+++ src/app/modules/charts/line-chart/line-chart.component.scss
@@ -1,9 +1,5 @@
 .container {
+  position: relative;
   width: 100%;
   height: 100%;
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
 }
