+++ src/app/modules/charts/line-chart/mock.ts
@@ -1,0 +1,7 @@
+export const MOCK_DATA = [
+  { date: '2019-01-04', value: 89},
+  { date: '2019-01-11', value: 96},
+  { date: '2019-01-18', value: 87},
+  { date: '2019-01-25', value: 99},
+  { date: '2019-02-04', value: 83}
+];
